refactor: 폴더이름 통일
