+++ src/pages/AlbumGroup/components/AlbumCard.scss
